# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mandate Hackathon API",
    "version": "0.1.0",
    "description": "Contract-first prototype: one trusted FastAPI transaction service, scoped agent access, independently retained audit checkpoints and a bounded verification component. HKD only. Core wallet rails are simulated. See README.md for normative transaction, delegation and evidence requirements. Optional commerce extension exposes local sandbox checkout separately from the local payment simulator; merchant orders remain synthetic."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local API; expose only scoped ingress when needed."
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/commerce/preview": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Policy Preview",
        "operationId": "policy_preview_commerce_preview_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/plans": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Basket Plans",
        "operationId": "basket_plans_commerce_plans_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/studies": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "List Studies",
        "operationId": "list_studies_commerce_studies_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      },
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Start Study",
        "operationId": "start_study_commerce_studies_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyStart"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/studies/{run_id}/finish": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Finish Study",
        "operationId": "finish_study_commerce_studies__run_id__finish_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyFinish"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/provider": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Provider Status",
        "operationId": "provider_status_commerce_provider_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/quotes/{quote_id}/approval": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Exact Approval",
        "operationId": "exact_approval_commerce_quotes__quote_id__approval_get",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/operations": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Operations",
        "operationId": "operations_commerce_operations_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      },
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Start Operation",
        "operationId": "start_operation_commerce_operations_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SandboxStart"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/operations/{operation_id}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Operation",
        "operationId": "operation_commerce_operations__operation_id__get",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/operations/{operation_id}/advance": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Advance Operation",
        "operationId": "advance_operation_commerce_operations__operation_id__advance_post",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/operations/{operation_id}/cancel": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Cancel Operation",
        "operationId": "cancel_operation_commerce_operations__operation_id__cancel_post",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/credentials/template": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Credential Template",
        "operationId": "credential_template_commerce_credentials_template_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialTemplate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    },
    "/commerce/credentials/verify": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Verify Credential",
        "operationId": "verify_credential_commerce_credentials_verify_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialVerify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-roles": [
          "user"
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server resolves an opaque configured credential to actor ID, role and authorized resources. Do not accept role/owner from request headers or body. Separate demo user and agent secrets; do not expose user secret to the agent."
      }
    },
    "schemas": {
...
      "Policy": {
        "properties": {
          "currency": {
            "type": "string",
            "const": "HKD",
            "title": "Currency"
          },
          "per_order_limit_minor": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Per Order Limit Minor"
          },
          "period_limits": {
            "items": {
              "$ref": "#/components/schemas/PeriodLimit"
            },
            "type": "array",
            "minItems": 1,
            "title": "Period Limits"
          },
          "allowed_merchant_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Allowed Merchant Ids"
          },
          "blocked_categories": {
            "items": {
              "type": "string",
              "enum": [
                "produce",
                "dairy",
                "eggs",
                "meat",
                "seafood",
                "bakery",
                "pantry",
                "beverage_non_alcoholic",
                "alcohol",
                "household",
                "unknown"
              ]
            },
            "type": "array",
            "title": "Blocked Categories"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "approval_above_minor": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Above Minor"
          },
          "velocity_limit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VelocityLimit"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_review": {
            "type": "boolean",
            "title": "Risk Review",
            "default": false
          },
          "web_purchases": {
            "type": "boolean",
            "title": "Web Purchases",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "currency",
          "per_order_limit_minor",
          "period_limits",
          "allowed_merchant_ids",
          "blocked_categories",
          "expires_at",
          "approval_above_minor"
        ],
        "title": "Policy"
      },
...
      "QuoteItemRequest": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quantity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "title": "QuoteItemRequest"
      },
...
      "CredentialTemplate": {
        "properties": {
          "mandate_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Mandate Id"
          },
          "public_x": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43,
            "title": "Public X"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mandate_id",
          "public_x"
        ],
        "title": "CredentialTemplate"
      },
      "CredentialVerify": {
        "properties": {
          "credential": {
            "type": "string",
            "maxLength": 30000,
            "minLength": 1,
            "title": "Credential"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "credential"
        ],
        "title": "CredentialVerify"
      },
      "PlanInput": {
        "properties": {
          "mandate_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Mandate Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuoteItemRequest"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mandate_id",
          "items"
        ],
        "title": "PlanInput"
      },
      "PreviewInput": {
        "properties": {
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "parent_mandate_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Mandate Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy"
        ],
        "title": "PreviewInput"
      },
      "StudyFinish": {
        "properties": {
          "actions": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Actions"
          },
          "errors": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Errors"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "ready",
              "failed",
              "abandoned"
            ],
            "title": "Outcome"
          },
          "source_url": {
            "type": "string",
            "maxLength": 1000,
            "title": "Source Url"
          },
          "observed_at": {
            "type": "string",
            "maxLength": 64,
            "title": "Observed At"
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "actions",
          "errors",
          "outcome",
          "source_url",
          "observed_at"
        ],
        "title": "StudyFinish"
      },
      "StudyStart": {
        "properties": {
          "participant": {
            "type": "string",
            "pattern": "^P[0-9]{2,4}$",
            "title": "Participant"
          },
          "task_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Task Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "manual",
              "agent"
            ],
            "title": "Mode"
          },
          "quote_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Quote Id"
          },
          "setup_seconds": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Setup Seconds"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "human",
              "scripted",
              "model",
              "fallback"
            ],
            "title": "Execution Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "participant",
          "task_id",
          "mode",
          "quote_id",
          "setup_seconds",
          "execution_mode"
        ],
        "title": "StudyStart"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SandboxStart": {
        "properties": {
          "mandate_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Mandate Id"
          },
          "quote_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Quote Id"
          },
          "approved_quote_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Approved Quote Hash"
          },
          "scenario": {
            "type": "string",
            "enum": [
              "happy",
              "order_failure",
              "lost_capture_response",
              "pending_refund",
              "failed_refund"
            ],
            "title": "Scenario",
            "default": "happy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mandate_id",
          "quote_id",
          "approved_quote_hash"
        ],
        "title": "SandboxStart"
      }
    }
  }
}